{{cookiecutter.project_slug}}/requirements/base.txt
pytz==2018.5  # https://github.com/stub42/pytz
python-slugify==1.2.6  # https://github.com/un33k/python-slugify
Pillow==5.2.0  # https://github.com/python-pillow/Pillow
{%- if cookiecutter.use_compressor == "y" %}
rcssmin==1.0.6{% if cookiecutter.windows == 'y' %} --install-option="--without-c-extensions"{% endif %}  # https://github.com/ndparker/rcssmin
{%- endif %}
argon2-cffi==18.3.0  # https://github.com/hynek/argon2_cffi
{%- if cookiecutter.use_whitenoise == 'y' %}
whitenoise==4.1  # https://github.com/evansd/whitenoise
{%- endif %}
redis>=2.10.5  # https://github.com/antirez/redis
{%- if cookiecutter.use_celery == "y" %}
celery==4.2.1  # pyup: <5.0  # https://github.com/celery/celery
{%- if cookiecutter.use_docker == 'y' %}
flower==0.9.2  # https://github.com/mher/flower
{%- endif %}
{%- endif %}

# Django
# ------------------------------------------------------------------------------
django==2.0.8  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.37.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
{%- if cookiecutter.use_compressor == "y" %}
django-compressor==2.2  # https://github.com/django-compressor/django-compressor
{%- endif %}
django-redis==4.9.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
djangorestframework-gis==0.13
djangorestframework-datatables==0.4.0
